requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-generativeai
httpx
python-dotenv
gunicorn
pandas==2.3.3
numpy==2.3.3
chromadb
sentence-transformers
pytest-asyncio